project/wcode-iai/easia-green/requirements-freeze.txt
albumentations==1.3.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
attrdict==2.0.1
attrs==25.3.0
babel==2.17.0
bce-python-sdk==0.9.29
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.12
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
Faker==18.9.0
fastapi==0.110.0
fastjsonschema==2.21.1
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-babel==4.0.0
fonttools==4.57.0
fqdn==1.5.1
fsspec==2025.3.2
future==1.0.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.37.0
imgaug==0.4.0
ipykernel==6.29.5
ipython==8.35.0
ipywidgets==8.1.6
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
labelImg==1.8.6
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.2
MarkupSafe==3.0.2
matplotlib==3.7.1
matplotlib-inline==0.1.7
mistune==3.1.3
more-itertools==10.6.0
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.0
notebook_shim==0.2.4
numpy==1.26.4
openai==1.75.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.5
opt-einsum==3.3.0
overrides==7.7.0
packaging==25.0
paddleocr==2.6.1.3
paddlepaddle==3.0.0
paddlepaddle-gpu==2.6.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
PassportEye==2.2.2
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer==20191125
pillow==10.2.0
platformdirs==4.3.7
premailer==3.10.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
protobuf==3.20.2
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyMuPDF==1.20.2
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
qudida==0.0.4
RapidFuzz==3.13.0
rarfile==4.2
referencing==0.36.2
regex==2023.12.25
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.36.3
sympy==1.13.1
termcolor==3.0.1
terminado==0.18.1
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.4.0
tomli==2.2.1
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tornado==6.4.2
tqdm==4.65.0
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.114
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.29.0
visualdl==2.5.3
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
